requirements.txt
arrow-odbc
connectorx
pandas
polars
pyarrow
python-dotenv
PyYAML
psutil
pyodbc
sqlalchemy